requirements.txt
streamlit
PyPDF2
openai
python-dotenv
supabase-py
gotrue
postgrest-py
pytest
realtime-py
requests